requirements.txt
django==4.0.0
djangorestframework==3.13.1
django-filter==21.1
Markdown==3.3.6
django-environ==0.8.1
PyMySQL==1.0.2
django-cors-headers==3.11.0
requests==2.27.1
attrs==21.4.0
drf-yasg==1.20.0
gevent==21.12.0
gunicorn==20.1.0